x86_64: fix register size errors when compiling on x86 mode

+++ libraries/x86_64/src/paging/cr3.rs
@@ -3,6 +3,8 @@ use essentials::address::PhysicalAddress;
 
 const ADDR_MASK: u64 = 0x_000f_ffff_ffff_f000;
 
+#[cfg(target_arch = "x86_64")]
+#[doc(cfg(target_arch = "x86_64"))]
 pub fn active_page() -> PhysicalAddress {
     let value: u64;
 
@@ -13,13 +15,17 @@ pub fn active_page() -> PhysicalAddress {
     PhysicalAddress::from(value & ADDR_MASK)
 }
 
+#[cfg(target_arch = "x86_64")]
+#[doc(cfg(target_arch = "x86_64"))]
 pub unsafe fn set_active_page(page_addr: PhysicalAddress) {
     let value = page_addr.as_u64() | (!ADDR_MASK);
     // clears out the add, without removing the flags.
     asm!("mov cr3, {}", in(reg) value, options(nostack, preserves_flags));
 }
 
+#[cfg(target_arch = "x86_64")]
+#[doc(cfg(target_arch = "x86_64"))]
 pub unsafe fn flush_page(page_addr: PhysicalAddress) {
-    let addr = page_addr.as_usize() as u64;
+    let addr = page_addr.as_usize();
     asm!("invlpg [{}]", in(reg) addr, options(nostack, preserves_flags));
 }
